+++ requirements.txt
@@ -1,0 +1,5 @@
+fastapi
+uvicorn
+python-dotenv
+twilio
+python-jose
